Extract RowSelectionItem into its own file and put in shared directory
