+++ app/src/main/java/com/example/alarmscratch/core/ui/shared/SailBoat.kt
@@ -1,4 +1,4 @@
-package com.example.alarmscratch.core.ui.core.component
+package com.example.alarmscratch.core.ui.shared
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.size
